requirements.txt
langchain
langchain_community
PyPDF2
streamlit==1.18.1
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
huggingface-hub==0.14.1
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2